pyproject.toml
[project]
name = "niimprint-server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "litestar[standard]>=2.20.0",
    "niimprint>=0.0.1",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pyserial>=3.5",
    "treepoem>=3.28.0",
]

[dependency-groups]
dev = [
    "dotenv>=0.9.9",
]
